Add files via upload
Here my question is whther a subclass can override the supercalss's private method or it is not visible in the child class? if I add  a private method with the same signature to the child class is it considered  a new method definition or an overriden version of the parent class?